added loading fallback for todos
