+++ src/app/components/Loading/styles.module.scss
@@ -1,0 +1,9 @@
+.container {
+  text-align: center;
+}
+
+.icon {
+  height: 1rem;
+  margin: auto 1.5rem auto auto;
+  color: #51646e;
+}
