+++ src/app/components/TodosList/styles.module.scss
@@ -1,0 +1,41 @@
+@import "../../../../global_styles/variables.scss";
+@import "../../../../global_styles/break-points.scss";
+
+.container {
+  padding: 1rem;
+  padding-top: $navbar-height + 1rem;
+
+  > ul {
+    list-style: none;
+    padding: 0;
+    display: grid;
+    gap: 1rem;
+
+    @include tablet {
+      grid-template-columns: 1fr;
+    }
+
+    @include laptop {
+      grid-template-columns: repeat(3, 1fr);
+    }
+
+    @include desktop {
+      grid-template-columns: repeat(4, 1fr);
+    }
+
+    @include desktop-large {
+      grid-template-columns: repeat(5, 1fr);
+    }
+  }
+
+  > p {
+    color: #464646;
+    margin: 0 1rem;
+    font-size: 0.9rem;
+  }
+}
+
+.no_todo_addded_text {
+  margin: 0 auto;
+  text-align: center;
+}
